Cloud/Chatbot/requirements.txt
# Function dependencies, for example:
# package>=version

Flask==1.0.2
dialogflow==0.5.1
google-api-core==1.4.1